fix(webui): move useMemo to actually-used view components
AgentsList and WorkspaceList have no callers; memoizing extraction there
had no effect. Move the useMemo to AgentsView and WorkspacesView, which
are the rendered components that call extractAgentsFromConversations and
extractWorkspacesFromConversations on every render.
