+++ webui/src/components/AgentsList.tsx
@@ -1,4 +1,3 @@
-import { useMemo } from 'react';
 import { Bot, ExternalLink, Plus, X } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import { extractAgentsFromConversations, type Agent } from '@/utils/workspaceUtils';
@@ -23,7 +22,7 @@ export const AgentsList: FC<AgentsListProps> = ({
 }) => {
   const { connectionConfig } = useApi();
   const baseUrl = connectionConfig.baseUrl.replace(/\/+$/, '');
-  const agents = useMemo(() => extractAgentsFromConversations(conversations), [conversations]);
+  const agents = extractAgentsFromConversations(conversations);
   const selectedAgent = use$(selectedAgent$);
 
   const handleSelectAgent = (agent: Agent) => {
