+++ webui/src/components/WorkspaceList.tsx
@@ -1,4 +1,3 @@
-import { useMemo } from 'react';
 import { Folder, X } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 import {
@@ -17,10 +16,7 @@ interface WorkspaceListProps {
 }
 
 export const WorkspaceList: FC<WorkspaceListProps> = ({ conversations }) => {
-  const workspaces = useMemo(
-    () => extractWorkspacesFromConversations(conversations),
-    [conversations]
-  );
+  const workspaces = extractWorkspacesFromConversations(conversations);
   const selectedWorkspace = use$(selectedWorkspace$);
 
   const handleSelectWorkspace = (workspace: WorkspaceProject) => {
